scenarious choosing by defining variables added

+++ ParallelMPISF/AgentOnNodeInfo.h
@@ -4,10 +4,10 @@ class AgentOnNodeInfo
 public:
 	AgentOnNodeInfo(void): isDeleted(false), _nodeID(-1), _agentID(-1) { };
 	bool isDeleted;
-	int _nodeID;
+	long long _nodeID;
 	long long _agentID;
-	AgentOnNodeInfo(int nodeID, int agentID) : isDeleted(false), _nodeID(nodeID), _agentID(agentID) { }
-	AgentOnNodeInfo(int nodeID, int agentID, bool isDeleted) : isDeleted(isDeleted), _nodeID(nodeID), _agentID(agentID) { }
+	AgentOnNodeInfo(int nodeID, long long agentID) : isDeleted(false), _nodeID(nodeID), _agentID(agentID) { }
+	AgentOnNodeInfo(int nodeID, long long agentID, bool isDeleted) : isDeleted(isDeleted), _nodeID(nodeID), _agentID(agentID) { }
 	~AgentOnNodeInfo(void);
 };
